ES6ify Client
ES6ify client.js and rename it to Client.js
