+++ src/server.js
@@ -2,7 +2,7 @@
 import _ from 'lodash';
 import assign from 'object-assign';
 import bcrypt from 'bcrypt-nodejs';
-import Client from './client';
+import Client from './Client';
 import ClientManager from './ClientManager';
 import express from 'express';
 import fs from 'fs';
